♻️ Refactor API module removing dependency to Domain
After more studies about feature wiring, it seems that is better to the
"-api" module does not depend on other modules. In this case, the
"Category-API" depended on Domain only due to the Mapper interface
defined there.

For now, I updated the code because I prefer some small duplication
rather than create unneeded dependency between modules. This update will
avoid "Category-API"  module to recompile if any change occur in the
Domain module, making the incremental builds faster as well.
