+++ features/category-api/build.gradle
@@ -2,8 +2,6 @@ apply from: "$rootDir/config/dependencies/feature_dependencies.gradle"
 apply plugin: "kotlin-parcelize"
 
 dependencies {
-    implementation project(":domain")
-
     implementation Deps.coroutines.core
     implementation Deps.koin.android
 }
