+++ features/category/src/main/java/com/escodro/category/di/CategoryModule.kt
@@ -1,10 +1,9 @@
 package com.escodro.category.di
 
-import com.escodro.category.mapper.CategoryMapperImpl
+import com.escodro.category.mapper.CategoryMapper
 import com.escodro.category.presentation.CategoryAddViewModel
 import com.escodro.category.presentation.CategoryEditViewModel
 import com.escodro.category.presentation.CategoryListViewModelImpl
-import com.escodro.categoryapi.mapper.CategoryMapper
 import com.escodro.categoryapi.presentation.CategoryListViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
@@ -17,5 +16,6 @@ val categoryModule = module {
     viewModel { CategoryAddViewModel(get(), get()) }
     viewModel { CategoryEditViewModel(get(), get(), get()) }
 
-    factory<CategoryMapper> { CategoryMapperImpl() }
+    // Mapper
+    factory { CategoryMapper() }
 }
